Rename file to reflect recent class name change.
