+++ python/lib/modeldata/dmod/modeldata/subset/__init__.py
@@ -1,3 +1,3 @@
 from .subset_definition import SubsetDefinition
-from .hf_subset_definition import HydrofabricSubset, SimpleHydrofabricSubset
+from .hydrofabric_subset import HydrofabricSubset, SimpleHydrofabricSubset
 from .subset_handler import SubsetHandler, SubsetHandlerImpl
